Add files via upload
Added a huggingface-based bloom filter usage notebook. Here is the reason why we did it: this enhancement will allow for data to be read on the fly from Hugging Face API without the user having to download the data to a local machine/server. Only the final annotated version will have to be saved to disk.  Also the Hugging Face API uses parquet file format.  Where as when one downloads Fineweb it could be of other formats like json etc.  So reading from Hugging Face API would make it very compatible.